+++ packages/shared/src/index.ts
@@ -40,5 +40,6 @@ export * from './sect-missions';
 export * from './sect-shop';
 export * from './sect-season';
 export * from './elemental';
+export * from './elemental-skills';
 export * from './territory';
 export * from './territory-buffs';